-rm reduntant conversion -encoding utf-8

+++ src/yass/YassAudioSample.java
@@ -126,11 +126,11 @@ public class YassAudioSample {
     public static byte[] getFrame(int frameNum) {
         int frameSize = sampleAudioFormat.getFrameSize();
 
-        int framePos = (int) ((frameNum - 1) * frameSize);
+        int framePos = (frameNum - 1) * frameSize;
         byte[] frame = new byte[frameSize];
         Byte sample = null;
         for (int i = 0; i < frameSize; i++) {
-            sample = (Byte) buffer.get(framePos + i + 1);
+            sample = buffer.get(framePos + i + 1);
             frame[i] = sample.byteValue();
         }
         return frame;
@@ -143,7 +143,7 @@ public class YassAudioSample {
      */
     public static int getLengthInFrames() {
         int frameLength = audioFormat.getFrameSize();
-        return (int) (buffer.size() / frameLength) - 1;
+        return buffer.size() / frameLength - 1;
     }
 
     /**
@@ -152,7 +152,7 @@ public class YassAudioSample {
     public static void playSample() {
         double soundFramePos = 1;
         int frameLength = sampleAudioFormat.getFrameSize();
-        int lengthInFrames = (int) (buffer.size() / frameLength) - 1;
+        int lengthInFrames = buffer.size() / frameLength - 1;
 
         byte[] sample = null;
